Forward declare hash specialisations and move implementations to numbers_templ.hpp
