+++ nano/lib/block_uniquer.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <nano/lib/numbers.hpp>
+#include <nano/lib/numbers_templ.hpp>
 #include <nano/lib/uniquer.hpp>
 
 namespace nano
